requirements.txt
anyio==4.4.0
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
dj-database-url==2.2.0
Django==4.0
django-cors==0.1
django-cors-headers==4.4.0
django-heroku==0.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
pytz==2024.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
Werkzeug==3.0.3
whitenoise==6.7.0
zipp==3.19.2
